backend/requirements.txt
boto3==1.28.85
requests==2.31.0
python-json-logger==2.0.7
PyJWT==2.8.0
